# perf: split validation logic for enforce and audit policies to return admission response earlier