--- a/pkg/webhooks/resource/handlers.go
+++ b/pkg/webhooks/resource/handlers.go
@@ -117,23 +117,21 @@ func (h *resourceHandlers) Validate(ctx context.Context, logger logr.Logger, req
 
 	logger.V(4).Info("processing policies for validate admission request", "validate", len(policies), "mutate", len(mutatePolicies), "generate", len(generatePolicies))
 
-	policyContext, err := h.buildPolicyContextFromAdmissionRequest(logger, request)
-	if err != nil {
-		return errorResponse(logger, request.UID, err, "failed create policy context")
-	}
-
-	vh := validation.NewValidationHandler(logger, h.kyvernoClient, h.engine, h.pCache, h.pcBuilder, h.eventGen, h.admissionReports, h.metricsConfig, h.configuration)
-
-	ok, msg, warnings := vh.HandleValidation(ctx, request, policies, policyContext, startTime)
+	vh := validation.NewValidationHandler(logger, h.kyvernoClient, h.engine, h.pCache, h.pcBuilder, h.eventGen, h.admissionReports, h.metricsConfig, h.configuration, h.nsLister)
+	ok, msg, warnings := vh.HandleValidationEnforce(ctx, request, policies, startTime)
+	go vh.HandleValidationAudit(ctx, request)
 	if !ok {
 		logger.Info("admission request denied")
 		return admissionutils.Response(request.UID, errors.New(msg), warnings...)
 	}
-	if !admissionutils.IsDryRun(request.AdmissionRequest) {
-		h.wg.Add(1)
-		go h.handleBackgroundApplies(ctx, logger, request, generatePolicies, mutatePolicies, startTime)
-		h.wg.Wait()
-	}
+
+	go func() {
+		if !admissionutils.IsDryRun(request.AdmissionRequest) {
+			h.wg.Add(1)
+			h.handleBackgroundApplies(ctx, logger, request, generatePolicies, mutatePolicies, startTime)
+			h.wg.Wait()
+		}
+	}()
 	return admissionutils.ResponseSuccess(request.UID, warnings...)
 }