--- a/pkg/webhooks/resource/validation/validation.go
+++ b/pkg/webhooks/resource/validation/validation.go
@@ -9,27 +9,29 @@ import (
 	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	"github.com/kyverno/kyverno/pkg/client/clientset/versioned"
 	"github.com/kyverno/kyverno/pkg/config"
-	"github.com/kyverno/kyverno/pkg/engine"
 	engineapi "github.com/kyverno/kyverno/pkg/engine/api"
+	"github.com/kyverno/kyverno/pkg/engine/policycontext"
 	"github.com/kyverno/kyverno/pkg/event"
 	"github.com/kyverno/kyverno/pkg/metrics"
 	"github.com/kyverno/kyverno/pkg/policycache"
 	"github.com/kyverno/kyverno/pkg/tracing"
 	admissionutils "github.com/kyverno/kyverno/pkg/utils/admission"
+	engineutils "github.com/kyverno/kyverno/pkg/utils/engine"
 	reportutils "github.com/kyverno/kyverno/pkg/utils/report"
 	"github.com/kyverno/kyverno/pkg/webhooks/handlers"
 	webhookutils "github.com/kyverno/kyverno/pkg/webhooks/utils"
 	"go.opentelemetry.io/otel/trace"
-	admissionv1 "k8s.io/api/admission/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	corev1listers "k8s.io/client-go/listers/core/v1"
 )
 
 type ValidationHandler interface {
 	// HandleValidation handles validating webhook admission request
 	// If there are no errors in validating rule we apply generation rules
 	// patchedResource is the (resource + patches) after applying mutation rules
-	HandleValidation(context.Context, handlers.AdmissionRequest, []kyvernov1.PolicyInterface, *engine.PolicyContext, time.Time) (bool, string, []string)
+	HandleValidationEnforce(context.Context, handlers.AdmissionRequest, []kyvernov1.PolicyInterface, time.Time) (bool, string, []string)
+	HandleValidationAudit(context.Context, handlers.AdmissionRequest)
 }
 
 func NewValidationHandler(
@@ -42,6 +44,7 @@ func NewValidationHandler(
 	admissionReports bool,
 	metrics metrics.MetricsConfigManager,
 	cfg config.Configuration,
+	nsLister corev1listers.NamespaceLister,
 ) ValidationHandler {
 	return &validationHandler{
 		log:              log,
@@ -53,6 +56,7 @@ func NewValidationHandler(
 		admissionReports: admissionReports,
 		metrics:          metrics,
 		cfg:              cfg,
+		nsLister:         nsLister,
 	}
 }
 
@@ -66,17 +70,27 @@ type validationHandler struct {
 	admissionReports bool
 	metrics          metrics.MetricsConfigManager
 	cfg              config.Configuration
+	nsLister         corev1listers.NamespaceLister
 }
 
-func (v *validationHandler) HandleValidation(
+func (v *validationHandler) HandleValidationEnforce(
 	ctx context.Context,
 	request handlers.AdmissionRequest,
 	policies []kyvernov1.PolicyInterface,
-	policyContext *engine.PolicyContext,
 	admissionRequestTimestamp time.Time,
 ) (bool, string, []string) {
 	resourceName := admissionutils.GetResourceName(request.AdmissionRequest)
 	logger := v.log.WithValues("action", "validate", "resource", resourceName, "operation", request.Operation, "gvk", request.Kind)
+
+	if len(policies) == 0 {
+		return true, "", nil
+	}
+
+	policyContext, err := v.buildPolicyContextFromAdmissionRequest(logger, request)
+	if err != nil {
+		return false, "failed create policy context", nil
+
+	}
 
 	var engineResponses []engineapi.EngineResponse
 	failurePolicy := kyvernov1.Ignore
@@ -94,7 +108,7 @@ func (v *validationHandler) HandleValidation(
 				engineResponse := v.engine.Validate(ctx, policyContext)
 				if engineResponse.IsNil() {
 					// we get an empty response if old and new resources created the same response
-					// allow updates if resource update doesnt change the policy evaluation
+					// allow updates if resource update doesn't change the policy evaluation
 					return
 				}
 
@@ -120,24 +134,56 @@ func (v *validationHandler) HandleValidation(
 		return false, webhookutils.GetBlockedMessages(engineResponses), nil
 	}
 
-	go v.handleAudit(ctx, policyContext.NewResource(), request, policyContext.NamespaceLabels(), engineResponses...)
+	go func() {
+		if needsReports(request, policyContext.NewResource(), v.admissionReports) {
+			if err := v.createReports(ctx, policyContext.NewResource(), request, engineResponses...); err != nil {
+				v.log.Error(err, "failed to create report")
+			}
+		}
+	}()
 
 	warnings := webhookutils.GetWarningMessages(engineResponses)
 	return true, "", warnings
 }
 
+func (v *validationHandler) HandleValidationAudit(
+	ctx context.Context,
+	request handlers.AdmissionRequest,
+) {
+	policyContext, err := v.buildPolicyContextFromAdmissionRequest(v.log, request)
+	if err != nil {
+		v.log.Error(err, "failed to build policy context")
+		return
+	}
+
+	needsReport := needsReports(request, policyContext.NewResource(), v.admissionReports)
+	tracing.Span(
+		context.Background(),
+		"",
+		fmt.Sprintf("AUDIT %s %s", request.Operation, request.Kind),
+		func(ctx context.Context, span trace.Span) {
+			responses, err := v.buildAuditResponses(ctx, request, policyContext)
+			if err != nil {
+				v.log.Error(err, "failed to build audit responses")
+			}
+			events := webhookutils.GenerateEvents(responses, false)
+			v.eventGen.Add(events...)
+			if needsReport {
+				v.createReports(ctx, policyContext.NewResource(), request, responses...)
+			}
+		},
+		trace.WithLinks(trace.LinkFromContext(ctx)),
+	)
+}
+
 func (v *validationHandler) buildAuditResponses(
 	ctx context.Context,
-	resource unstructured.Unstructured,
 	request handlers.AdmissionRequest,
-	namespaceLabels map[string]string,
+	policyContext *policycontext.PolicyContext,
 ) ([]engineapi.EngineResponse, error) {
 	gvr := schema.GroupVersionResource(request.Resource)
 	policies := v.pCache.GetPolicies(policycache.ValidateAudit, gvr, request.SubResource, request.Namespace)
-	policyContext, err := v.pcBuilder.Build(request.AdmissionRequest, request.Roles, request.ClusterRoles, request.GroupVersionKind)
-	if err != nil {
-		return nil, err
-	}
+
 	var responses []engineapi.EngineResponse
 	for _, policy := range policies {
 		tracing.ChildSpan(
@@ -145,7 +191,7 @@ func (v *validationHandler) buildAuditResponses(
 			"pkg/webhooks/resource/validate",
 			fmt.Sprintf("POLICY %s/%s", policy.GetNamespace(), policy.GetName()),
 			func(ctx context.Context, span trace.Span) {
-				policyContext := policyContext.WithPolicy(policy).WithNamespaceLabels(namespaceLabels)
+				policyContext := policyContext.WithPolicy(policy)
 				response := v.engine.Validate(ctx, policyContext)
 				responses = append(responses, response)
 			},
@@ -154,51 +200,31 @@ func (v *validationHandler) buildAuditResponses(
 	return responses, nil
 }
 
-func (v *validationHandler) handleAudit(
+func (v *validationHandler) buildPolicyContextFromAdmissionRequest(logger logr.Logger, request handlers.AdmissionRequest) (*policycontext.PolicyContext, error) {
+	policyContext, err := v.pcBuilder.Build(request.AdmissionRequest, request.Roles, request.ClusterRoles, request.GroupVersionKind)
+	if err != nil {
+		return nil, err
+	}
+	namespaceLabels := make(map[string]string)
+	if request.Kind.Kind != "Namespace" && request.Namespace != "" {
+		namespaceLabels = engineutils.GetNamespaceSelectorsFromNamespaceLister(request.Kind.Kind, request.Namespace, v.nsLister, logger)
+	}
+	policyContext = policyContext.WithNamespaceLabels(namespaceLabels)
+	return policyContext, nil
+}
+
+func (v *validationHandler) createReports(
 	ctx context.Context,
 	resource unstructured.Unstructured,
 	request handlers.AdmissionRequest,
-	namespaceLabels map[string]string,
 	engineResponses ...engineapi.EngineResponse,
-) {
-	createReport := v.admissionReports
-	if admissionutils.IsDryRun(request.AdmissionRequest) {
-		createReport = false
+) error {
+	report := reportutils.BuildAdmissionReport(resource, request.AdmissionRequest, engineResponses...)
+	if len(report.GetResults()) > 0 {
+		_, err := reportutils.CreateReport(ctx, report, v.kyvernoClient)
+		if err != nil {
+			return err
+		}
 	}
-	// we don't need reports for deletions
-	if request.Operation == admissionv1.Delete {
-		createReport = false
-	}
-	// check if the resource supports reporting
-	if !reportutils.IsGvkSupported(schema.GroupVersionKind(request.Kind)) {
-		createReport = false
-	}
-	// if the underlying resource has no UID don't create a report
-	if resource.GetUID() == "" {
-		createReport = false
-	}
-	tracing.Span(
-		context.Background(),
-		"",
-		fmt.Sprintf("AUDIT %s %s", request.Operation, request.Kind),
-		func(ctx context.Context, span trace.Span) {
-			responses, err := v.buildAuditResponses(ctx, resource, request, namespaceLabels)
-			if err != nil {
-				v.log.Error(err, "failed to build audit responses")
-			}
-			events := webhookutils.GenerateEvents(responses, false)
-			v.eventGen.Add(events...)
-			if createReport {
-				responses = append(responses, engineResponses...)
-				report := reportutils.BuildAdmissionReport(resource, request.AdmissionRequest, responses...)
-				if len(report.GetResults()) > 0 {
-					_, err = reportutils.CreateReport(ctx, report, v.kyvernoClient)
-					if err != nil {
-						v.log.Error(err, "failed to create report")
-					}
-				}
-			}
-		},
-		trace.WithLinks(trace.LinkFromContext(ctx)),
-	)
+	return nil
 }